Add method Regex::getNumCaptures()
This returns the actual number of captures in the pattern.

This is useful for APIs that pass captures as variadic arguments to a
callback: PCRE2 only returns captures up to the last one that matches, so if
some at the end do not match, the number of captures passed can vary.

This can confuse the callee, which would typically be expecting one argument
per capture.

The particular reason I made this change was for the Node binding of PCRE2,
which copies the JavaScript API String.prototype.replace(). Callback
functions for this API have the following signature:

f(match, capture1, ..., capturen, offset, string, named_groups)

It’s bad enough trying to deal with a variable number of capture arguments
in JavaScript; it’s even worse in TypeScript!