Allow getNumberedSubstrings to leave unmatched captures null
If a capture does not match, it should not be filled in with an empty
string, that’s not the same thing!

Also, fix a bug where captures that matched at position 0 were not returned.

Add an option macro JPCRE2_UNSET_CAPTURES_NULL, which requires C++17 (add
machinery for this) to use std::optional so that NumSub always contains the
right number of elements, but some may be absent.

Update test_issue_29.cpp to test that captures are filled in if and only if
they are expected from the input data.

Do not set CPPFLAGS in configure.ac, so that the user can set it, for
example to set JPCRE2_UNSET_CAPTURES_NULL.